pyproject.toml
[project]
name = "sleepstack"
version = "0.1.0"
description = "Binaural beats + ambience mixer for narration-ready sleep tracks"
readme = "README.md"
license = {text = "MIT"}
authors = [
    { name = "Tyler Kozlowski", email = "Tyler.J.Kozlowski@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "numpy",
    "yt-dlp>=2025.9.26",
]

[project.optional-dependencies]
dev = ["pytest", "black", "mypy"]

[project.scripts]
sleepstack = "sleepstack.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sleepstack"]
include = [
    "assets/**/*",
]

# Dev tool configs
[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
